project/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Backend application"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
soupsieve = "2.8"
beautifulsoup4 = "4.13.5"
cloudscraper = "1.2.71"
naked = "0.1.32"
shellescape = "3.8.1"
requests-toolbelt = "1.0.0"
tonsdk = "1.0.15"
tonutils = "0.5.3"
tonviewer = "1.1.3"
cachetools = "5.5.2"
slowapi = "0.1.9"
fastapi-cache2 = "0.2.2"
redis = "4.6.0"
# MySQL drivers (replaced with PostgreSQL)
# asyncmy = "0.2.10"
# pymysql = "1.1.0"
# aiomysql = "0.2.0"

# PostgreSQL drivers
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"
aiofiles = "24.1.0"
aiogram = "3.22.0"
aiohappyeyeballs = "2.6.1"
aiohttp = "3.12.2"
aiosignal = "1.4.0"
aiosqlite = "0.21.0"
alembic = "1.16.5"
annotated-types = "0.7.0"
anyio = "4.11.0"
attrs = "25.3.0"
certifi = "2025.8.3"
cffi = "2.0.0"
click = "8.3.0"
colorama = "0.4.6"
cryptography = "45.0.7"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "^4.1.2"
curl-cffi = "0.13.0"
dnspython = "2.8.0"
email-validator = "2.3.0"
fastapi = "0.118.0"
fastapi-cli = "0.0.13"
fastapi-cloud-cli = "0.2.1"
frozenlist = "1.7.0"
greenlet = "3.2.4"
h11 = "0.16.0"
httpcore = "1.0.9"
httptools = "0.6.4"
httpx = "0.28.1"
idna = "3.10"
jinja2 = "3.1.6"
magic-filter = "1.0.12"
mako = "1.3.10"
markdown-it-py = "4.0.0"
markupsafe = "3.0.3"
mdurl = "0.1.2"
multidict = "6.6.4"
opentele = "1.15.1"
propcache = "0.3.2"
pyaes = "1.6.1"
pyasn1 = "0.6.1"
pycparser = "2.23"
pydantic = "2.11.9"
pydantic-settings = "2.11.0"
pydantic-core = "2.33.2"
pygments = "2.19.2"
pyqt5 = "5.15.11"
pyqt5-qt5 = "5.15.2"
pyqt5-sip = "12.17.0"
python-dotenv = "1.1.1"
python-multipart = "0.0.20"
pyyaml = "6.0.3"
rich = "14.1.0"
rich-toolkit = "0.15.1"
rignore = "0.6.4"
rsa = "4.9.1"
sentry-sdk = "2.39.0"
shellingham = "1.5.4"
sniffio = "1.3.1"
sqlalchemy = "2.0.43"
starlette = "0.48.0"
telegram-webapp-auth = "3.0.6"
telethon = "1.41.2"
tgcrypto = "1.2.5"
typer = "0.19.2"
typing-inspection = "0.4.1"
typing-extensions = "4.15.0"
urllib3 = "2.5.0"
uvicorn = "0.37.0"
watchfiles = "1.1.0"
websockets = "15.0.1"
yarl = "1.20.1"
fake-useragent = "1.5.1"
pycryptodome = "3.21.0"
loguru = "0.7.2"
psycopg = "^3.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.1.0"
black = "^23.7.0"
flake8 = "^6.1.0"
mypy = "^1.5.0"
locust = "^2.20.0"
ruff = "^0.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
# Длина строки
line-length = 120
target-version = "py310"

# Исключаемые директории
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "build",
    "dist",
    ".pytest_cache",
    ".mypy_cache",
    "logs",
]

[tool.ruff.lint]
# Включаем правила
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort (сортировка импортов)
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "SIM",    # flake8-simplify
    "RUF",    # Ruff-specific rules
]

# Игнорируем некоторые правила
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "B904",   # raise from
    "RUF001", # ambiguous unicode characters (для русских комментариев)
    "RUF002", # ambiguous unicode characters in docstrings
    "RUF003", # ambiguous unicode characters in comments
]

# Автофикс для безопасных правил
fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.isort]
# Настройки сортировки импортов
known-first-party = ["app", "project"]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]
lines-after-imports = 2

[tool.ruff.lint.per-file-ignores]
# Игнорируем некоторые правила для тестов
"tests/**/*.py" = ["S101", "B007", "B017"]  # assert allowed in tests, unused loop vars, pytest.raises(Exception)
# Игнорируем star imports в __init__.py
"**/__init__.py" = ["F401", "F403", "F405"]
# Игнорируем undefined names в моделях (SQLAlchemy forward references)
"**/models/*.py" = ["F821"]
# Игнорируем в скриптах
"scripts/**/*.py" = ["B007"]

[tool.ruff.format]
# Настройки форматирования
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Постепенно включим
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true
plugins = ["pydantic.mypy"]

# Игнорируем сторонние библиотеки без типов
[[tool.mypy.overrides]]
module = [
    "telethon.*",
    "aiogram.*",
    "opentele.*",
    "tonsdk.*",
    "tonutils.*",
    "tonviewer.*",
    "cloudscraper.*",
]
ignore_missing_imports = true

# Строгая типизация для наших модулей
[[tool.mypy.overrides]]
module = "app.api.schemas.*"
disallow_untyped_defs = true
disallow_any_generics = true

[[tool.mypy.overrides]]
module = "app.use_cases.*"
disallow_untyped_defs = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
